Sign Up test refactoring.
