+++ SignUpTests.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
 <suite name="All Test Suite">
-    <test verbose="2" preserve-order="true" name="C:/Users/LiubaIlchenko/GITCl/GreenCity/GreenCityTests">
+    <test verbose="2" preserve-order="true" name="C:/Users/LiubaIlchenko/GITCl/GreenCity/GreenCityTests"  parallel="classes" thread-count="20">
         <classes>
             <class name="com.softserve.edu.greencity.ui.tests.RegistrationTests"/>
             <class name="com.softserve.edu.greencity.ui.tests.RegisterPageTests"/>
