+++ src/test/java/com/softserve/edu/greencity/ui/pages/common/TopUserComponent.java
@@ -46,7 +46,7 @@ public class TopUserComponent {
 	}
 
 	public String getUserNameButtonText() {
-		return getUserNameButton().getText().trim();
+		return getUserNameButton().getText();
 	}
 
 	public void clickUserNameButton() {
