+++ src/test/java/com/softserve/edu/greencity/ui/tests/RegisterPageTests.java
@@ -15,6 +15,222 @@ import org.testng.annotations.Test;
 
 public class RegisterPageTests extends GreenCityTestRunner {
 
+    @DataProvider
+    public Object[][] validUserCredentials() {
+        return new Object[][]{
+                {UserRepository.get().defaultUserCredentials()},};
+    }
+
+    @DataProvider
+    public Object[][] emptyFields() {
+        return new Object[][]{
+                {UserRepository.get().emptyUserCredentials()},
+        };
+    }
+
+    @DataProvider
+    public Object[][] invalidFields() {
+        return new Object[][]{
+                {UserRepository.get().invalidUserCredentials()},};
+    }
+
+    @DataProvider
+    public Object[][] invalidEmail() {
+        return new Object[][]{
+                {UserRepository.get().invalidEmailUserCredentials()},};
+    }
+
+    @DataProvider
+    public Object[][] invalidPassUppercaseUserCreds() {
+        return new Object[][]{
+                {UserRepository.get().invalidPassUppercaseUserCreds()},};
+    }
+
+    @DataProvider
+    public Object[][] invalidPassDigitUserCreds() {
+        return new Object[][]{
+                {UserRepository.get().invalidPassDigitUserCreds()},};
+    }
+
+    @DataProvider
+    public Object[][] invalidPassLowercaseUserCreds() {
+        return new Object[][]{
+                {UserRepository.get().invalidPassLowercaseUserCreds()},};
+    }
+
+    @DataProvider
+    public Object[][] invalidPassSpecCharUserCreds() {
+        return new Object[][]{
+                {UserRepository.get().invalidPassSpecCharUserCreds()},};
+    }
+
+    @DataProvider
+    public Object[][] invalidPassLengthUserCreds() {
+        return new Object[][]{
+                {UserRepository.get().invalidPassLengthUserCreds()},};
+    }
+
+    @DataProvider
+    public Object[][] invalidPassSpaceUserCreds() {
+        return new Object[][]{
+                {UserRepository.get().invalidPassSpaceUserCreds()},};
+    }
+
+    @DataProvider
+    public Object[][] invalidNameCredentials() {
+        return new Object[][]{
+                {UserRepository.get().invalidNameCredentials()},};
+    }
+
+    @DataProvider
+    public Object[][] invalidConfirmPass() {
+        return new Object[][]{
+                {UserRepository.get().invalidConfirmPassCredentials()},};
+    }
+    /**
+     * Filling all the fields on the Register page without registering and
+     * switch to Login page.
+     */
+    @Test(dataProvider = "validUserCredentials")
+    public void checkIfSignUpButtonEnabled(User userLoginCredentials) {
+        loadApplication();
+        logger.info("Starting checkIfSignUpButtonEnabled. Input values = "
+                + userLoginCredentials.toString());
+
+        logger.info("Click on Sign up button");
+        RegisterComponent registerComponent = new TopGuestComponent(driver).clickSignUpLink();
+
+        logger.info("Get a title text of the modal window: "
+                + registerComponent.getTitleString());
+
+        Assert.assertEquals("Hello!", registerComponent.getTitleString(),
+                "This is not a register modal:(");
+        ManualRegisterComponent manualRegisterComponent = new ManualRegisterComponent(driver);
+
+        Assert.assertTrue(manualRegisterComponent.signUpIsDisabled());
+
+        logger.info(
+                "Filling out the fields with valid credentials without clicking on Sign up button");
+        manualRegisterComponent.fillFieldsWithoutRegistration(userLoginCredentials);
+
+        Assert.assertFalse(manualRegisterComponent.signUpIsDisabled());
+
+    }
+
+    /**
+     * Opening register modal window, reading its title,
+     * switching to login modal window, reading its title
+     * to make sure the transition has been executed.
+     */
+    @Test
+    public void navigateFromSignUpToSignIn() {
+        loadApplication();
+        logger.info("Starting navigateFromSignUpToSignIn");
+
+        logger.info("Click on Sign up button");
+        RegisterComponent registerComponent = new TopGuestComponent(driver).clickSignUpLink();
+
+        logger.info("Get a title text of the modal window: "
+                + registerComponent.getTitleString());
+
+        Assert.assertEquals("Hello!", registerComponent.getTitleString(),
+                "This is not a register modal:(");
+
+        Assert.assertEquals("Please enter your details to sign up", registerComponent.getSubtitleString(),
+                "This is not a register modal:(");
+
+        logger.info("Click on Sign in button");
+        LoginComponent loginComponent = registerComponent.clickSignInLink();
+
+        Assert.assertEquals("Welcome back!", loginComponent.getTitleString(),
+                "This is not a login modal:(");
+
+        Assert.assertEquals("Please enter your details to sign in", loginComponent.getSubtitleString(),
+                "This is not a login modal:(");
+    }
+
+    /**
+     * Putting empty values into register form
+     * and reading the validation messages.
+     */
+    @Test(dataProvider = "emptyFields", description = "GC-502, GC-207, GC-208, GC-209, GC-210")
+    public void checkEmptyFieldsValidation(User userLoginCredentials) {
+        loadApplication();
+        logger.info("Starting checkEmptyFieldsValidation. Input values = "
+                + userLoginCredentials.toString());
+
+        logger.info("Click on Sign up button");
+        RegisterComponent registerComponent = new TopGuestComponent(driver).clickSignUpLink();
+
+        logger.info("Get a title text of the modal window: "
+                + registerComponent.getTitleString());
+
+        Assert.assertEquals("Hello!", registerComponent.getTitleString(),
+                "This is not a register modal:(");
+
+        ManualRegisterComponent manualRegisterComponent = registerComponent.getManualRegisterComponent();
+
+        logger.info("Enter empty values into the form");
+        manualRegisterComponent.registrationWrongUser(userLoginCredentials);
+
+        Assert.assertEquals(manualRegisterComponent.getEmailValidatorText(),
+                "Email is required",
+                "The validation message is not equal to the expected one");
+
+        Assert.assertEquals(manualRegisterComponent.getUserNameValidatorText(),
+                "User name is required",
+                "The validation message is not equal to the expected one");
+
+        Assert.assertEquals(manualRegisterComponent.getPasswordValidatorText(),
+                "Password is required",
+                "The validation message is not equal to the expected one");
+
+        Assert.assertEquals(manualRegisterComponent.getPasswordConfirmValidatorText(),
+                "Password is required",
+                "The validation message is not equal to the expected one");
+
+    }
+
+    /**
+     * Putting invalid values into register form
+     * and reading the validation messages.
+     */
+    @Test(dataProvider = "invalidFields")
+    public void checkInvalidFieldsValidation(User userLoginCredentials) {
+        loadApplication();
+        logger.info("Starting checkInvalidFieldsValidation. Input values = "
+                + userLoginCredentials.toString());
+
+        logger.info("Click on Sign up button");
+        RegisterComponent registerComponent = new TopGuestComponent(driver).clickSignUpLink();
+
+        logger.info("Get a title text of the modal window: "
+                + registerComponent.getTitleString());
+
+        Assert.assertEquals("Hello!", registerComponent.getTitleString(),
+                "This is not a register modal:(");
+
+        ManualRegisterComponent manualRegisterComponent = registerComponent.getManualRegisterComponent();
+
+        logger.info("Enter invalid values into the form");
+        manualRegisterComponent.registrationWrongUser(userLoginCredentials);
+
+        Assert.assertEquals(manualRegisterComponent.getEmailValidatorText(),
+                "Please check that your e-mail address is indicated correctly",
+                "The validation message is not equal to the expected one");
+
+        Assert.assertEquals(manualRegisterComponent.getPasswordValidatorText(),
+                "Password must be at least 8 characters in length",
+                "The validation message is not equal to the expected one");
+
+        Assert.assertEquals(manualRegisterComponent.getPasswordConfirmValidatorText(),
+                "Password has contain at least one character of Uppercase letter (A-Z), " +
+                        "Lowercase letter (a-z), Digit (0-9), Special character (~`!@#$%^&*()+=_-{}[]|:;”’?/<>,.)",
+                "The validation message is not equal to the expected one");
+
+    }
+
+
     @Test(description = "GC-501")
     public void navigateFromSignInToSignUp() {
 
@@ -57,7 +273,7 @@ public class RegisterPageTests extends GreenCityTestRunner {
                 .clickSignUpLink()
                 .getManualRegisterComponent();
 
-        logger.info("Enter invalid email and valid name and password into the form: ");
+        logger.info("Enter invalid email and valid name and password into the form");
         manualRegisterComponent.fillFieldsWithoutRegistration(invalidEmailCredentials);
 
         Assert.assertEquals(manualRegisterComponent.getEmailValidatorText(),
@@ -68,11 +284,6 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
     }
 
-    @DataProvider
-    public Object[][] invalidEmail() {
-        return new Object[][]{
-                {UserRepository.get().invalidEmailUserCredentials()},};
-    }
 
     @Test(dataProvider = "invalidConfirmPass", description = "GC-519")
     public void invalidConfirmPassRegistration(User userLoginCredentials) {
@@ -91,7 +302,7 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
         ManualRegisterComponent manualRegisterComponent = registerComponent.getManualRegisterComponent();
 
-        logger.info("Enter invalid values into the form: ");
+        logger.info("Enter invalid values into the form ");
         manualRegisterComponent.fillFieldsWithoutRegistration(userLoginCredentials);
 
         Assert.assertEquals(manualRegisterComponent.getPasswordConfirmValidatorText(),
@@ -100,12 +311,6 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
         Assert.assertTrue(manualRegisterComponent.signUpIsDisabled());
 
-    }
-
-    @DataProvider
-    public Object[][] invalidConfirmPass() {
-        return new Object[][]{
-                {UserRepository.get().invalidConfirmPassCredentials()},};
     }
 
 
@@ -126,7 +331,7 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
         ManualRegisterComponent manualRegisterComponent = registerComponent.getManualRegisterComponent();
 
-        logger.info("Enter invalid values into the form: ");
+        logger.info("Enter invalid values into the form ");
         manualRegisterComponent.fillFieldsWithoutRegistration(userLoginCredentials);
 
         Assert.assertEquals(manualRegisterComponent.getPasswordValidatorText(),
@@ -138,11 +343,6 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
     }
 
-    @DataProvider
-    public Object[][] invalidPassUppercaseUserCreds() {
-        return new Object[][]{
-                {UserRepository.get().invalidPassUppercaseUserCreds()},};
-    }
 
     @Test(dataProvider = "invalidPassDigitUserCreds", description = "GC-517")
     public void invalidPassDigitValidation(User userLoginCredentials) {
@@ -161,7 +361,7 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
         ManualRegisterComponent manualRegisterComponent = registerComponent.getManualRegisterComponent();
 
-        logger.info("Enter invalid values into the form: ");
+        logger.info("Enter invalid values into the form");
         manualRegisterComponent.fillFieldsWithoutRegistration(userLoginCredentials);
 
         Assert.assertEquals(manualRegisterComponent.getPasswordValidatorText(),
@@ -171,12 +371,6 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
         Assert.assertTrue(manualRegisterComponent.signUpIsDisabled());
 
-    }
-
-    @DataProvider
-    public Object[][] invalidPassDigitUserCreds() {
-        return new Object[][]{
-                {UserRepository.get().invalidPassDigitUserCreds()},};
     }
 
     @Test(dataProvider = "invalidPassLowercaseUserCreds", description = "GC-517")
@@ -196,7 +390,7 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
         ManualRegisterComponent manualRegisterComponent = registerComponent.getManualRegisterComponent();
 
-        logger.info("Enter invalid values into the form: ");
+        logger.info("Enter invalid values into the form");
         manualRegisterComponent.fillFieldsWithoutRegistration(userLoginCredentials);
 
         Assert.assertEquals(manualRegisterComponent.getPasswordValidatorText(),
@@ -208,11 +402,6 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
     }
 
-    @DataProvider
-    public Object[][] invalidPassLowercaseUserCreds() {
-        return new Object[][]{
-                {UserRepository.get().invalidPassLowercaseUserCreds()},};
-    }
 
     @Test(dataProvider = "invalidPassSpecCharUserCreds", description = "GC-517")
     public void invalidPassSpecCharValidation(User userLoginCredentials) {
@@ -231,7 +420,7 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
         ManualRegisterComponent manualRegisterComponent = registerComponent.getManualRegisterComponent();
 
-        logger.info("Enter invalid values into the form: ");
+        logger.info("Enter invalid values into the form ");
         manualRegisterComponent.fillFieldsWithoutRegistration(userLoginCredentials);
 
 
@@ -245,13 +434,8 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
     }
 
-    @DataProvider
-    public Object[][] invalidPassSpecCharUserCreds() {
-        return new Object[][]{
-                {UserRepository.get().invalidPassSpecCharUserCreds()},};
-    }
 
-    //GC-198
+
     @Test(dataProvider = "invalidPassLengthUserCreds", description = "GC-198, GC-517")
     public void invalidPassLengthValidation(User userLoginCredentials) {
         loadApplication();
@@ -269,7 +453,7 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
         ManualRegisterComponent manualRegisterComponent = registerComponent.getManualRegisterComponent();
 
-        logger.info("Enter invalid values into the form: ");
+        logger.info("Enter invalid values into the form ");
         manualRegisterComponent.fillFieldsWithoutRegistration(userLoginCredentials);
 
         Assert.assertEquals(manualRegisterComponent.getPasswordValidatorText(),
@@ -280,11 +464,6 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
     }
 
-    @DataProvider
-    public Object[][] invalidPassLengthUserCreds() {
-        return new Object[][]{
-                {UserRepository.get().invalidPassLengthUserCreds()},};
-    }
 
     @Test(dataProvider = "invalidPassSpaceUserCreds", description = "GC-517")
     public void invalidPassSpaceValidation(User userLoginCredentials) {
@@ -303,25 +482,18 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
         ManualRegisterComponent manualRegisterComponent = registerComponent.getManualRegisterComponent();
 
-        logger.info("Enter invalid values into the form: ");
+        logger.info("Enter invalid values into the form ");
         manualRegisterComponent.fillFieldsWithoutRegistration(userLoginCredentials);
 
         Assert.assertEquals(manualRegisterComponent.getPasswordValidatorText(),
                 "Password must be at least 8 characters long without spaces",
                 "The validation message is not equal to the expected one");
 
-        Assert.assertFalse(manualRegisterComponent.signUpIsDisabled());
+        Assert.assertTrue(manualRegisterComponent.signUpIsDisabled());
 
     }
 
-    @DataProvider
-    public Object[][] invalidPassSpaceUserCreds() {
-        return new Object[][]{
-                {UserRepository.get().invalidPassSpaceUserCreds()},};
-    }
-
-
-    @Test(description = "GC-1214")
+    @Test(description = "GC-499")
     public void checkCloseRegisterModalButton() {
         loadApplication();
         logger.info("Starting checkCloseRegisterModalButton:");
@@ -364,7 +536,7 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
         ManualRegisterComponent manualRegisterComponent = registerComponent.getManualRegisterComponent();
 
-        logger.info("Enter invalid values into the form: ");
+        logger.info("Enter invalid values into the form ");
         manualRegisterComponent.fillPasswordFieldPassHidden("Valid!1");
 
         Assert.assertEquals(manualRegisterComponent.getPasswordField().getAttribute("type"),
@@ -375,15 +547,6 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
     }
 
-
-    //GC-204
-    //Verify that Email must be existence and unique while new user registration
-    //GC-200
-    //Verify that unregistered user sees popup window 'Sign up' after clicking on the “My habits” button
-    //GC-203
-    //Verify that User is redirected to My habits as a Registered User after he has entered valid credentials
-   //GC-216
-    // Verify 'Sign up' page UI
 
     @Test(description = "GC-485")
     public void checkBackgroundIsDimmed() {
@@ -399,7 +562,7 @@ public class RegisterPageTests extends GreenCityTestRunner {
     }
 
 
-    @Test(dataProvider = "invalidNameCredentials", description = "GC-205")
+        @Test(dataProvider = "invalidNameCredentials", description = "GC-205")
     public void checkUserFieldMaxLength(User userLoginCredentials) {
         loadApplication();
         logger.info("Starting checkInvalidFieldsValidation. Input values = "
@@ -417,7 +580,7 @@ public class RegisterPageTests extends GreenCityTestRunner {
 
         ManualRegisterComponent manualRegisterComponent = registerComponent.getManualRegisterComponent();
 
-        logger.info("Entering values into the form: ");
+        logger.info("Entering values into the form ");
         manualRegisterComponent.registrationWrongUser(userLoginCredentials);
 
         String userFieldValue = manualRegisterComponent.getUserNameField().getAttribute("value");
@@ -427,12 +590,16 @@ public class RegisterPageTests extends GreenCityTestRunner {
                 "The invalid string is not concatenated");
 
     }
-    @DataProvider
-    public Object[][] invalidNameCredentials() {
-        return new Object[][]{
-                {UserRepository.get().invalidNameCredentials()},};
-    }
 
+
+    //GC-204
+    //Verify that Email must be existence and unique while new user registration
+    //GC-200
+    //Verify that unregistered user sees popup window 'Sign up' after clicking on the “My habits” button
+    //GC-203
+    //Verify that User is redirected to My habits as a Registered User after he has entered valid credentials
+    //GC-216
+    // Verify 'Sign up' page UI
     //GC-487
     //Verify UI of the Registration form on different screen resolutions
