+++ src/test/java/com/softserve/edu/greencity/ui/tools/CookiesAndStorageHelper.java
@@ -41,6 +41,7 @@ public class CookiesAndStorageHelper {
         webStorage.getLocalStorage().clear();
 
 
+
     }
 
     public void cleanGMailCookiesAndStorages() {
